requirements.txt
# OCR Serial Version Requirements (No Ray)
fastapi>=0.115.12
uvicorn>=0.24.0
pydantic>=2.0.0
flask>=3.1.1
gunicorn>=23.0.0
imgaug>=0.4.0
lmdb>=1.6.2

# ONNX Runtime - use one of these options:
# For GPU support (requires CUDA)
# onnxruntime-gpu>=1.15.0,<1.20.0

# For CPU-only (more stable, recommended for initial testing)
onnxruntime>=1.15.0,<1.20.0

opencv-contrib-python>=4.7.0.72
opencv-python-headless>=4.7.0.72
pyclipper>=1.3.0.post6
requests>=2.32.3
scikit-image>=0.25.2
shapely>=2.1.1
tqdm>=4.67.1
numpy>=1.21.0,<2.0.0
pillow>=8.0.0,<11.0.0
